rename: add a package and move files
