+++ src/main/java/gift/api/Application.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
