+++ src/main/java/gift/api/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.api;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
