+++ src/main/java/gift/api/ProductDao.java
@@ -1,8 +1,7 @@
-package gift;
+package gift.api;
 
 import java.sql.Types;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.simple.JdbcClient;
 import org.springframework.stereotype.Repository;
